+++ src/mod_runtime_api/mod.rs
@@ -1,2 +1,3 @@
 pub mod entities;
 pub mod runtime_api;
+pub mod errors;
